Move MenuSpecHelper to Spec::Support::MenuHelper
